#16: ...but actually, keep ast.go as top-level thing
