+++ fishi/fishi.go
@@ -20,7 +20,7 @@ import (
 )
 
 type Results struct {
-	AST  *ir.AST
+	AST  *AST
 	Tree *types.ParseTree
 }
 
@@ -127,12 +127,14 @@ func Parse(source []byte, opts Options) (Results, error) {
 
 	r := Results{}
 	// now, try to make a parse tree
-	ast, pt, err := fishiFront.AnalyzeString(string(preprocessedSource))
+	nodes, pt, err := fishiFront.AnalyzeString(string(preprocessedSource))
 	r.Tree = pt // need to do this before we return
 	if err != nil {
 		return r, err
 	}
-	r.AST = &ast
+	r.AST = &AST{
+		Nodes: nodes,
+	}
 
 	return r, nil
 }
@@ -140,7 +142,7 @@ func Parse(source []byte, opts Options) (Results, error) {
 // GetFrontend gets the frontend for the fishi compiler-compiler. If cffFile is
 // provided, it is used to load the cached parser from disk. Otherwise, a new
 // frontend is created.
-func GetFrontend(opts Options) (ictiobus.Frontend[ir.AST], error) {
+func GetFrontend(opts Options) (ictiobus.Frontend[[]ir.ASTBlock], error) {
 	// check for preload
 	var preloadedParser ictiobus.Parser
 	if opts.ParserCFF != "" && opts.ReadCache {
@@ -150,7 +152,7 @@ func GetFrontend(opts Options) (ictiobus.Frontend[ir.AST], error) {
 			if errors.Is(err, os.ErrNotExist) {
 				preloadedParser = nil
 			} else {
-				return ictiobus.Frontend[ir.AST]{}, fmt.Errorf("loading cachefile %q: %w", opts.ParserCFF, err)
+				return ictiobus.Frontend[[]ir.ASTBlock]{}, fmt.Errorf("loading cachefile %q: %w", opts.ParserCFF, err)
 			}
 		}
 	}
@@ -160,7 +162,7 @@ func GetFrontend(opts Options) (ictiobus.Frontend[ir.AST], error) {
 		ParserTrace: opts.ParserTrace,
 	}
 
-	fishiFront := fe.Frontend[ir.AST](ir.HooksTable, feOpts, preloadedParser)
+	fishiFront := fe.Frontend[[]ir.ASTBlock](ir.HooksTable, feOpts, preloadedParser)
 
 	// check the parser encoding if we generated a new one:
 	if preloadedParser == nil && opts.ParserCFF != "" && opts.WriteCache {
@@ -185,7 +187,7 @@ func GetFrontend(opts Options) (ictiobus.Frontend[ir.AST], error) {
 
 		sddErr := fishiFront.SDT.Validate(fishiFront.Parser.Grammar(), fishiFront.IRAttribute, di, valProd)
 		if sddErr != nil {
-			return ictiobus.Frontend[ir.AST]{}, fmt.Errorf("sdd validation error: %w", sddErr)
+			return ictiobus.Frontend[[]ir.ASTBlock]{}, fmt.Errorf("sdd validation error: %w", sddErr)
 		}
 	}
